# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: GameStore API
  description: Homework on Swagger||Game Store
  version: 1.0.0

paths:
  /game-store:
    get:
      summary: Get all games
      responses:
        '200':
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'

  /game-store/{id}:
    get:
      summary: Get game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found

    put:
      summary: Update game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameUpdate'
      responses:
        '200':
          description: Game updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Game not found

    delete:
      summary: Delete game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game deleted
          content:
            text/plain:
              schema:
                type: string
                example: "Game deleted with id : 1"
        '404':
          description: Game not found

  /game-store/add:
    post:
      summary: Add a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameInput'
      responses:
        '200':
          description: Game added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'

components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: StarDew Valley
        genre:
          type: string
          example: Farm RPG
      required:
        - id
        - title
        - genre

    GameInput:
      type: object
      properties:
        title:
          type: string
          example: StarDew Valley
        genre:
          type: string
          example: Farm RPG
      required:
        - title
        - genre

    GameUpdate:
      type: object
      properties:
        title:
          type: string
          example: StarDew Valley Updated
        genre:
          type: string
          example: Farm, RPG, Fighting
      required:
        - title
        - genre